code review and data mapping cleanup
